Split cig.go to multiple files to make it cleaner
